Created Dashboard 2
Created sheet 'state-wise confirmed cases in India'  which gives overall confirmed cases of each state in India in decreasing order, sheet 'statistics for chosen state' which gives bar graph of a chosen state from the dropdown of cured, confirmed , deaths and sheet 'comparison of cured vs death for chosen state' which gives a state-wise graph of death vs cured percentage where the formula is used. Created a Dashboard showing state-wise analysis for a selected state in a drop down.